+++ packages/@tailwindcss-browser/src/assets.ts
@@ -1,0 +1,11 @@
+import index from 'tailwindcss/index.css'
+import preflight from 'tailwindcss/preflight.css'
+import theme from 'tailwindcss/theme.css'
+import utilities from 'tailwindcss/utilities.css'
+
+export const css = {
+  index,
+  preflight,
+  theme,
+  utilities,
+}